Moved chevron up icon to common
